+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Version 0.0.5 (2019-04-17)
 
 * bugfix: FRIN to be on input
 * enhancement: add an example
+* bugfix: node T is now included
+* bugfix: nodes FRIN/FROUT/T now have separate IDs
 
 
 Version 0.0.4 (2019-04-1{5,6,7})